+++ src/utils.js
@@ -83,44 +83,6 @@ const humanToMultiaddrBuffer = (str, { strict = true } = {}) => {
   return multiaddrBuffer;
 };
 
-const minBn = (bnArray) => {
-  let min = new BN(bnArray[0]);
-  bnArray.map((e) => {
-    if (e.lt(min)) min = e;
-    return min;
-  });
-  return min;
-};
-
-const isEthAddress = (address, { strict = false } = {}) => {
-  const isHexString = typeof address === 'string' && address.substr(0, 2) === '0x';
-  const isAddress = isHexString && address.length === 42;
-  if (!isAddress && strict) {
-    throw Error(`Address ${address} is not a valid Ethereum address`);
-  }
-  return isAddress;
-};
-
-const isBytes32 = (str, { strict = false } = {}) => {
-  if (
-    typeof str !== 'string'
-    || str.length !== 66
-    || str.substr(0, 2) !== '0x'
-  ) {
-    if (strict) throw new Error(`${str} is not a valid Bytes32 HexString`);
-    return false;
-  }
-  return true;
-};
-
-const isString = (str, { strict = true } = {}) => {
-  if (typeof str !== 'string') {
-    if (strict) throw new Error(`${str} is not a string`);
-    return false;
-  }
-  return true;
-};
-
 const cleanRPC = (rpcObj) => {
   const keys = Object.keys(rpcObj);
   const cleanObj = keys.reduce((accu, curr) => {
@@ -319,18 +281,12 @@ const throwIfMissing = () => {
   throw Error('Missing parameter');
 };
 
-const ensureString = val => String(val);
-
 module.exports = {
   BN,
-  isString,
-  isEthAddress,
   checksummedAddress,
-  isBytes32,
   cleanRPC,
   checkEvent,
   getEventFromLogs,
-  minBn,
   bnToEthersBn,
   ethersBnToBn,
   ethersBigNumberify: bigNumberify,
@@ -347,7 +303,6 @@ module.exports = {
   getSalt,
   NULL_ADDRESS,
   NULL_BYTES32,
-  ensureString,
   throwIfMissing,
   signTypedDatav3,
   truncateBnWeiToBnNRlc,